Rename panel files and add Debug mode
