--- src/ThreeEditor/js/sidebar/object/Objects.js
@@ -1,12 +1,0 @@
-export * from './Object.Beam';
-export * from './Object.CSG';
-export * from './Object.Differentials';
-export * from './Object.Dimensions';
-export * from './Object.Filter';
-export * from './Object.Grid';
-export * from './Object.Info';
-export * from './Object.Material';
-export * from './Object.Placement';
-export * from './Object.Quantity';
-export * from './Object.Settings';
-export * from './Object.ZoneCalculate';
