+++ src/ThreeEditor/js/sidebar/panel/Panel.Abstract.ts
