+++ src/ThreeEditor/js/sidebar/panel/Panel.Beam.ts
@@ -8,7 +8,7 @@ import {
 } from '../../../util/Ui/Uis';
 import { Editor } from '../../Editor';
 import { UINumber, UIRow, UIText } from '../../libs/ui';
-import { ObjectAbstract } from './Object.Abstract';
+import { ObjectAbstract } from './Panel.Abstract';
 
 export class ObjectBeam extends ObjectAbstract {
 	object?: Beam;
