+++ src/ThreeEditor/js/sidebar/panel/Panel.Quantity.ts
@@ -1,22 +1,16 @@
-import { ScoringOutput } from '../../../util/Scoring/ScoringOutput';
 import * as Scoring from '../../../util/Scoring/ScoringOutputTypes';
 import { ScoringQuantity } from '../../../util/Scoring/ScoringQuantity';
 import {
-	createFullwidthButton,
 	createRowConditionalNumber,
 	createRowConditionalSelect,
 	createRowSelect,
 	hideUIElement,
 	showUIElement
 } from '../../../util/Ui/Uis';
-import {
-	AddQuantityCommand,
-	SetOutputSettingsCommand,
-	SetQuantityValueCommand
-} from '../../commands/Commands';
+import { SetQuantityValueCommand } from '../../commands/Commands';
 import { Editor } from '../../Editor';
-import { UIButton, UICheckbox, UINumber, UIPanel, UIRow, UISelect, UIText } from '../../libs/ui';
-import { ObjectAbstract } from './Object.Abstract';
+import { UICheckbox, UINumber, UIRow, UISelect } from '../../libs/ui';
+import { ObjectAbstract } from './Panel.Abstract';
 
 export class ObjectQuantity extends ObjectAbstract {
 	object?: ScoringQuantity;
