+++ src/ThreeEditor/js/sidebar/panel/Panels.js
@@ -1,0 +1,12 @@
+export * from './Panel.Beam';
+export * from './Panel.CSG';
+export * from './Panel.Differentials';
+export * from './Panel.Dimensions';
+export * from './Panel.Filter';
+export * from './Panel.Grid';
+export * from './Panel.Info';
+export * from './Panel.Material';
+export * from './Panel.Placement';
+export * from './Panel.Quantity';
+export * from './Panel.Settings';
+export * from './Panel.ZoneCalculate';
