+++ src/util/Config.ts
@@ -1,2 +1,3 @@
 export const BACKEND_URL = process.env.REACT_APP_BACKEND_URL ?? 'http://localhost:5000';
 export const DEMO_MODE = process.env.REACT_APP_TARGET === 'demo';
+export const DEBUG_MODE = process.env.REACT_APP_TARGET === 'debug';
